+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 - The available types of working directories for the current environment are now shown in the output of `qq submit -h`.
 - Fixed a regression from v0.5: missing size property in `qq nodes` is now correctly intepreted as zero size.
 
+### Internal changes
+- `Wiper.delete` method has been renamed to `Wiper.wipe`.
+- `Killer.terminate` method has been renamed to `Killer.kill`.
+
 ***
 
 ## Version 0.5.1